+++ package/Source~/source/Providers/ReferencePointProvider.h
@@ -10,11 +10,18 @@
 class ReferencePointProvider : public IUnityXRReferencePointProvider
 {
 public:
+    ReferencePointProvider();
+    virtual ~ReferencePointProvider();
     virtual bool UNITY_INTERFACE_API TryAddReferencePoint(const UnityXRPose& xrPose, UnityXRTrackableId& outId, UnityXRTrackingState& outTrackingState) override;
     virtual bool UNITY_INTERFACE_API TryRemoveReferencePoint(const UnityXRTrackableId& id) override;
     virtual bool UNITY_INTERFACE_API GetAllReferencePoints(IUnityXRReferencePointAllocator& allocator) override;
 
+    static ReferencePointProvider* Get() { return s_Instance; }
+
+    UnityXRTrackableId AttachReferencePoint(UnityXRTrackableId trackableId, UnityXRPose pose);
+
 private:
+    static ReferencePointProvider* s_Instance;
     typedef std::map<UnityXRTrackableId, WrappedAnchor> IdToAnchorMap;
     IdToAnchorMap m_IdToAnchorMap;
 };